Better handling of errors
We don't need to check the cURL error number as this is covered by the
response. A non-FALSE response means we definitely have something back
from the web service (i.e. the request has been completed), then the
status code needs to be examined to identify whether the email was sent
or not.